fixes issue with unused components
